# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: KinaKeep SME Bookkeeping API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /businesses/{businessId}/suppliers:
    get:
      tags: [Suppliers]
      operationId: listSuppliers
      summary: List suppliers for a business
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Supplier'
    post:
      tags: [Suppliers]
      operationId: createSupplier
      summary: Create a supplier
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierInput'
      responses:
        '201':
          description: Supplier created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
  /businesses/{businessId}/suppliers/{supplierId}:
    get:
      tags: [Suppliers]
      operationId: getSupplier
      summary: Get a supplier
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - $ref: '#/components/parameters/SupplierIdParam'
      responses:
        '200':
          description: Supplier details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
    patch:
      tags: [Suppliers]
      operationId: updateSupplier
      summary: Update a supplier
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - $ref: '#/components/parameters/SupplierIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplierUpdate'
      responses:
        '200':
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Supplier'
    delete:
      tags: [Suppliers]
      operationId: deleteSupplier
      summary: Delete a supplier
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - $ref: '#/components/parameters/SupplierIdParam'
      responses:
        '204':
          description: Deleted

  # ─── Documents ────────────────────────────────────────────────────────────
...
  /businesses/{businessId}/reports/summary:
    get:
      tags: [Reports]
      operationId: getReportSummary
      summary: Get report summary for a date range
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Report summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportSummary'
  /businesses/{businessId}/reports/income-breakdown:
    get:
      tags: [Reports]
      operationId: getIncomeBreakdown
      summary: Get income breakdown by category
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Income breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryBreakdown'
  /businesses/{businessId}/reports/expense-breakdown:
    get:
      tags: [Reports]
      operationId: getReportExpenseBreakdown
      summary: Get expense breakdown by category for report
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Expense breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryBreakdown'
  /businesses/{businessId}/reports/missing-receipts:
    get:
      tags: [Reports]
      operationId: getMissingReceiptsReport
      summary: Get all transactions with missing receipts
      parameters:
        - $ref: '#/components/parameters/BusinessIdParam'
        - name: startDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Transactions with missing receipts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
components:
  parameters:
...
    BusinessIdParam:
      name: businessId
      in: path
      required: true
      schema:
        type: number
...
    SupplierIdParam:
      name: supplierId
      in: path
      required: true
      schema:
        type: number
...
  schemas:
...
    Transaction:
      type: object
      required: [id, businessId, type, amount, description, date, paymentMethod, referenceNumber, receiptStatus, createdAt]
      properties:
        id:
          type: number
        businessId:
          type: number
        type:
          type: string
          enum: [income, expense]
        amount:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        categoryId:
          type: [integer, 'null']
        categoryName:
          type: [string, 'null']
        paymentMethod:
          type: string
        customerId:
          type: [integer, 'null']
        customerName:
          type: [string, 'null']
        supplierId:
          type: [integer, 'null']
        supplierName:
          type: [string, 'null']
        referenceNumber:
          type: string
        receiptPath:
          type: [string, 'null']
        receiptStatus:
          type: string
          enum: [attached, missing]
        receiptReference:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Supplier:
      type: object
      required: [id, businessId, name, totalPurchases, createdAt]
      properties:
        id:
          type: number
        businessId:
          type: number
        name:
          type: string
        contactPerson:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        email:
          type: [string, 'null']
        address:
          type: [string, 'null']
        notes:
          type: [string, 'null']
        totalPurchases:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SupplierInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
    SupplierUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        contactPerson:
          type: [string, 'null']
        phone:
          type: [string, 'null']
        email:
          type: [string, 'null']
        address:
          type: [string, 'null']
        notes:
          type: [string, 'null']

    # ─── Document ──────────────────────────────────────────────────────────
...
    CategoryBreakdown:
      type: object
      required: [categoryName, total, count]
      properties:
        categoryName:
          type: string
        total:
          type: number
        count:
          type: number

    # ─── Reports ───────────────────────────────────────────────────────────
    ReportSummary:
      type: object
      required: [startDate, endDate, totalMoneyIn, totalMoneyOut, netCashFlow, transactionCount, receiptsAttached, receiptsMissing, receiptCoverage]
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalMoneyIn:
          type: string
        totalMoneyOut:
          type: string
        netCashFlow:
          type: string
        transactionCount:
          type: number
        receiptsAttached:
          type: number
        receiptsMissing:
          type: number
        receiptCoverage:
          type: number